formatting updates

+++ gpaCalcPython/flaskr/schema.sql
@@ -12,6 +12,6 @@ CREATE TABLE class (
   author_id INTEGER NOT NULL,
   classname TEXT NOT NULL,
   grade INTEGER NOT NULL,
-  credits INTEGER NOT NULL,
+  credits REAL NOT NULL,
   FOREIGN KEY (author_id) REFERENCES user (id)
 );
